+++ main/inc/mlx_config.h
@@ -30,8 +30,6 @@
 # define KEY_PLUS	24
 # define KEY_MINUS	27
 
-
-
 typedef struct	s_img
 {
 	void	*img;
@@ -54,9 +52,6 @@ typedef struct    s_mlx_confix
 
 void    mlx_my_putpixel(t_img *data, int x, int y, int color);
 void    mlx_my_init(t_mlx_confix *vars);
-int     mlx_my_close(int key, t_mlx_confix *vars);
 void    fill_image_with_color(t_mlx_confix *vars);
-
-
 
 #endif
