+++ main/srcs/base_object/color_debug.c
@@ -2,6 +2,6 @@
 
 int     print_color(t_color *color)
 {
-    printf("color: %d, %d, %d\n", color->r, color->g, color->b);
+    printf("color: [%d, %d, %d]\n", color->r, color->g, color->b);
     return (0);
 }
