+++ main/srcs/utils/my_math.c
@@ -1,4 +1,4 @@
-#include "../inc/minirt.h"
+#include "../../inc/minirt.h"
 
 // return min exist positive value
 double solve_quadratic(double a, double b, double c)
@@ -23,3 +23,4 @@ double solve_quadratic(double a, double b, double c)
     }
     return (-1);
 }
+
